Make it possible to get error outside of tx
Some box error members do not have accessor functions, so a workaround
is used: evaluating Lua code to get offsets for "private" fields.
However, it is dangerous to use such code outside of the tx thread.
Therefore, remove the Lua evaluation and not obtain the error members
which Tarantool does not provide an accessor for.